refactor: separate config tabs and set history as a tab
